Safari-611.1.10/LayoutTests/canvas/philip/tests/2d.clearRect.negative.html

<!DOCTYPE html>
<title>Canvas test: 2d.clearRect.negative</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="clear-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.clearRect(0, 0, 50, 25);
ctx.clearRect(100, 0, -50, 25);
ctx.clearRect(0, 50, 50, -25);
ctx.clearRect(100, 50, -50, -25);
_assertPixel(canvas, 25,12, 0,0,0,0, "25,12", "0,0,0,0");
_assertPixel(canvas, 75,12, 0,0,0,0, "75,12", "0,0,0,0");
_assertPixel(canvas, 25,37, 0,0,0,0, "25,37", "0,0,0,0");
_assertPixel(canvas, 75,37, 0,0,0,0, "75,37", "0,0,0,0");


});
</script>
