LayoutTests/canvas/philip/tests/2d.clearRect.negative-expected.txt

< [index] >
2d.clearRect.negative
clearRect of negative sizes works
References: 2d.clearRect
Actual output:
Expected output:

Passed
